Update lab03-canvas-app.md
The imported TabControl does not have an orange background. This should be corrected. Adding the step to set Fill to Organ sometime causes tab text to not be visible.

Suggestion: It is more efficient to change the tab SelectedColor to WhiteSmoke before duplicating the screen.

There is a mismatch between the provided JavaScript and the naming of the New Reports Screen.